rename Node class to Variant class and move to haplotypes module
