+++ happler/tree/treebuilder.py
@@ -1,4 +1,5 @@
-from .tree import Node, Haplotype, Tree
+from .tree import Tree
+from .haplotype import Variant, Haplotype
 from ..data import Genotypes, Phenotypes
 
 
@@ -35,13 +36,13 @@ class TreeBuilder:
             index into :py:attr:`~.TreeBuilder.gens`
         """
         # step one: initialize the tree
-        root_node = Node.from_np(self.gens[root], idx=root)
+        root_node = Variant.from_np(self.gens[root], idx=root)
         self.tree = Tree(root_node)
         # step two: create the tree
         self._create_tree(root_node)
 
     def _create_tree(
-        self, parent: Node, parent_hap: Haplotype = None, parent_idx: int = 0
+        self, parent: Variant, parent_hap: Haplotype = None, parent_idx: int = 0
     ):
         """
         Recursive helper to the run() function
@@ -50,7 +51,7 @@ class TreeBuilder:
 
         Parameters
         ----------
-        parent : Node
+        parent : Variant
             An existing node in the tree under which we should consider creating a subtree
         parent_hap : Haplotype
             The haplotype containing all variants up to (but NOT including) the parent
@@ -69,7 +70,7 @@ class TreeBuilder:
             parent_hap.append(parent, allele)
             self._create_tree(best_variant, parent_hap, new_node_idx)
 
-    def _find_split(self, parent: Haplotype, parent_idx: int, allele: int) -> Node:
+    def _find_split(self, parent: Haplotype, parent_idx: int, allele: int) -> Variant:
         """
         Find the variant that best fits under the parent_idx node with the allele edge
 
@@ -77,4 +78,6 @@ class TreeBuilder:
         ----------
         TODO
         """
-        pass
+        # step 1: transform GT matrix into haplotype matrix
+        # step 2: test assoc
+        # step 3:
